Remove NetworkObject and NetworkService; keep a NetworkXObjectConfig for the systemAdapter for now
