+++ pkg/pillar/cmd/zedrouter/strongswan.go
@@ -25,231 +25,6 @@ const (
 	AnyIpAddr         = "%any"
 )
 
-func strongswanCreate(ctx *zedrouterContext, config types.NetworkServiceConfig,
-	status *types.NetworkServiceStatus) error {
-
-	log.Infof("strongswanCreate(%s)\n", config.DisplayName)
-
-	// parse and structure the config
-	vpnConfig, err := strongSwanConfigGet(ctx, config)
-	if err != nil {
-		return err
-	}
-
-	// stringify and store in status
-	bytes, err := json.Marshal(vpnConfig)
-	if err != nil {
-		return err
-	}
-
-	status.OpaqueStatus = string(bytes)
-	log.Infof("StrongSwanCreate: %s\n", status.OpaqueStatus)
-
-	// reset any previous config
-	if err := ipSecServiceInactivate(vpnConfig); err != nil {
-		return err
-	}
-
-	// create the ipsec config files, tunnel, routes  and filter-rules
-	if err := strongSwanVpnCreate(vpnConfig); err != nil {
-		log.Errorf("%s StrongSwanVpn create\n", err.Error())
-		return err
-	}
-	return nil
-}
-
-func strongswanDelete(status *types.NetworkServiceStatus) {
-
-	log.Infof("strongswanDelete(%s)\n", status.DisplayName)
-
-	vpnConfig, err := strongSwanVpnStatusParse(status.OpaqueStatus)
-	if err != nil {
-		log.Infof("strongswanDelete config absent\n")
-		return
-	}
-
-	if err := strongSwanVpnDelete(vpnConfig); err != nil {
-		log.Errorf("%s StrongSwanVpn delete\n", err.Error())
-	}
-}
-
-func strongswanActivate(ctx *zedrouterContext, config types.NetworkServiceConfig,
-	status *types.NetworkServiceStatus, netstatus *types.NetworkObjectStatus) error {
-
-	log.Infof("strongswanActivate(%s)\n", status.DisplayName)
-
-	vpnConfig, err := strongSwanVpnStatusParse(status.OpaqueStatus)
-	if err != nil {
-		log.Errorf("StrongSwanVpn config absent\n")
-		return err
-	}
-
-	if err := strongSwanVpnActivate(vpnConfig); err != nil {
-		log.Errorf("%s StrongSwanVpn activate\n", err.Error())
-		return err
-	}
-	return nil
-}
-
-func strongswanInactivate(ctx *zedrouterContext, status *types.NetworkServiceStatus,
-	netstatus *types.NetworkObjectStatus) {
-
-	log.Infof("strongswanInactivate(%s)\n", status.DisplayName)
-	vpnConfig, err := strongSwanVpnStatusParse(status.OpaqueStatus)
-	if err != nil {
-		log.Errorf("StrongSwanVpn config absent\n")
-	}
-
-	if err := strongSwanVpnInactivate(vpnConfig); err != nil {
-		log.Errorf("%s StrongSwanVpn inactivate\n", err.Error())
-		return
-	}
-}
-
-// XXX remove this when we remove NetworkObject
-// StrongSwan Vpn IpSec Tenneling handler routines
-func strongSwanConfigGet(ctx *zedrouterContext,
-	config types.NetworkServiceConfig) (types.VpnServiceConfig, error) {
-
-	port := types.NetLinkConfig{}
-	appLink := types.NetLinkConfig{}
-	vpnConfig := types.VpnServiceConfig{}
-	appNetPresent := false
-
-	// if adapter is not set, return
-	if config.Adapter == "" {
-		return vpnConfig, errors.New("port config is absent")
-	}
-
-	// port ip address error
-	srcIp, err := types.GetLocalAddrAnyNoLinkLocal(*ctx.deviceNetworkStatus, 0,
-		config.Adapter)
-	if err != nil {
-		return vpnConfig, err
-	}
-
-	port.Name = types.AdapterToIfName(ctx.deviceNetworkStatus,
-		config.Adapter)
-	port.IpAddr = srcIp.String()
-
-	// app net information
-	// XXX whole function will be removed when we remove NetworkObject
-	appNet := lookupNetworkObjectConfig(ctx, config.AppLink.String())
-	if appNet != nil {
-		if appNet.Type != types.NT_IPV4 {
-			return vpnConfig, errors.New("appnet is not IPv4")
-		}
-		appNetPresent = true
-		appLink.SubnetBlock = appNet.Subnet.String()
-	}
-
-	vpnCloudConfig, err := strongSwanVpnConfigParse(config.OpaqueConfig)
-	if err != nil {
-		return vpnConfig, err
-	}
-
-	// XXX:TBD  host names to ip addresses in the configuration
-	vpnConfig.VpnRole = vpnCloudConfig.VpnRole
-	vpnConfig.IsClient = vpnCloudConfig.IsClient
-	vpnConfig.PolicyBased = vpnCloudConfig.PolicyBased
-	vpnConfig.GatewayConfig = vpnCloudConfig.GatewayConfig
-	vpnConfig.PortConfig = port
-	vpnConfig.AppLinkConfig = appLink
-
-	// fill and validate the ip address/subnet
-	if vpnConfig.GatewayConfig.IpAddr == PortIpAddrType {
-		vpnConfig.GatewayConfig.IpAddr = port.IpAddr
-	}
-	if appNetPresent &&
-		vpnConfig.GatewayConfig.SubnetBlock == AppLinkSubnetType {
-		vpnConfig.GatewayConfig.SubnetBlock = appLink.SubnetBlock
-	}
-	if err := vpnValidateIpAddr(vpnConfig.GatewayConfig.IpAddr, true); err != nil {
-		return vpnConfig, err
-	}
-	if err := vpnValidateSubnet(vpnConfig.GatewayConfig.SubnetBlock); err != nil {
-		return vpnConfig, err
-	}
-	vpnConfig.ClientConfigList = make([]types.VpnClientConfig,
-		len(vpnCloudConfig.ClientConfigList))
-
-	for idx, ssClientConfig := range vpnCloudConfig.ClientConfigList {
-		clientConfig := new(types.VpnClientConfig)
-		clientConfig.IpAddr = ssClientConfig.IpAddr
-		clientConfig.SubnetBlock = ssClientConfig.SubnetBlock
-		clientConfig.PreSharedKey = ssClientConfig.PreSharedKey
-		clientConfig.TunnelConfig.Name = fmt.Sprintf("%s_%d", vpnConfig.VpnRole, idx)
-		clientConfig.TunnelConfig.Key = "100"
-		clientConfig.TunnelConfig.Mtu = "1419"
-		clientConfig.TunnelConfig.Metric = "50"
-		clientConfig.TunnelConfig.LocalIpAddr = ssClientConfig.TunnelConfig.LocalIpAddr
-		clientConfig.TunnelConfig.RemoteIpAddr = ssClientConfig.TunnelConfig.RemoteIpAddr
-
-		if clientConfig.IpAddr == PortIpAddrType {
-			clientConfig.IpAddr = port.IpAddr
-		}
-		if appNetPresent &&
-			clientConfig.SubnetBlock == AppLinkSubnetType {
-			clientConfig.SubnetBlock = appLink.SubnetBlock
-		}
-		if clientConfig.IpAddr == "" {
-			clientConfig.IpAddr = AnyIpAddr
-		}
-		// validate the ip address/subnet values
-		if err := vpnValidateIpAddr(clientConfig.IpAddr, false); err != nil {
-			return vpnConfig, err
-		}
-		if err := vpnValidateSubnet(clientConfig.SubnetBlock); err != nil {
-			return vpnConfig, err
-		}
-		tunnelConfig := clientConfig.TunnelConfig
-		if err := vpnValidateLinkLocal(tunnelConfig.LocalIpAddr); err != nil {
-			return vpnConfig, err
-		}
-		if err := vpnValidateLinkLocal(tunnelConfig.RemoteIpAddr); err != nil {
-			return vpnConfig, err
-		}
-		if clientConfig.SubnetBlock == vpnConfig.GatewayConfig.SubnetBlock {
-			return vpnConfig, errors.New("Peer is on Same Subnet")
-		}
-		vpnConfig.ClientConfigList[idx] = *clientConfig
-	}
-
-	if !vpnConfig.IsClient {
-		if vpnConfig.GatewayConfig.IpAddr != port.IpAddr {
-			errorStr := vpnConfig.GatewayConfig.IpAddr
-			errorStr = errorStr + ", port: " + port.IpAddr
-			return vpnConfig, errors.New("IpAddr Mismatch, GatewayIp: " + errorStr)
-		}
-		// ensure appNet match
-		if appNetPresent &&
-			vpnConfig.GatewayConfig.SubnetBlock != "" &&
-			vpnConfig.GatewayConfig.SubnetBlock != appLink.SubnetBlock {
-			errorStr := vpnConfig.GatewayConfig.SubnetBlock + ", appNet: " + appLink.SubnetBlock
-			return vpnConfig, errors.New("Subnet Mismatch: " + errorStr)
-		}
-	} else {
-		// for clients
-		if appNetPresent {
-			for _, clientConfig := range vpnConfig.ClientConfigList {
-				if clientConfig.SubnetBlock != "" &&
-					clientConfig.SubnetBlock != appLink.SubnetBlock {
-					errorStr := clientConfig.SubnetBlock
-					errorStr = errorStr + ", appNet: " + appLink.SubnetBlock
-					return vpnConfig, errors.New("Subnet Mismatch: " + errorStr)
-				}
-			}
-		}
-	}
-	if log.GetLevel() == log.DebugLevel {
-		if bytes, err := json.Marshal(vpnConfig); err == nil {
-			log.Debugf("strongSwanVpnConfigGet(): %s\n",
-				string(bytes))
-		}
-	}
-	return vpnConfig, nil
-}
-
 func strongSwanVpnConfigParse(opaqueConfig string) (types.VpnServiceConfig, error) {
 	log.Infof("strongSwanVpnConfigParse(): parsing %s\n", opaqueConfig)
 	vpnConfig := types.VpnServiceConfig{}
@@ -742,34 +517,6 @@ func vpnValidateIpAddr(ipAddrStr string, isValid bool) error {
 	return nil
 }
 
-func strongSwanVpnStatusGet(ctx *zedrouterContext,
-	status *types.NetworkServiceStatus) bool {
-	change := false
-	vpnConfig, err := strongSwanVpnStatusParse(status.OpaqueStatus)
-	if err != nil {
-		log.Infof("StrongSwanVpn config absent\n")
-		return change
-	}
-	vpnStatus := new(types.ServiceVpnStatus)
-	if err := ipSecStatusCmdGet(vpnStatus); err != nil {
-		return change
-	}
-	if err := swanCtlCmdGet(vpnStatus); err != nil {
-		return change
-	}
-
-	vpnStatus.PolicyBased = vpnConfig.PolicyBased
-
-	// if tunnel state have changed, update
-	if change = isVpnStatusChanged(status.VpnStatus, vpnStatus); change {
-		log.Debugf("vpn state change:%v\n", vpnStatus)
-		status.VpnStatus = vpnStatus
-	}
-	// push the vpnMetrics here
-	publishVpnMetrics(ctx, status, vpnStatus)
-	return change
-}
-
 func isVpnStatusChanged(oldStatus, newStatus *types.ServiceVpnStatus) bool {
 	if oldStatus == nil {
 		return true
@@ -853,32 +600,6 @@ func matchConnState(oldConn, newConn *types.VpnConnStatus) bool {
 	return true
 }
 
-func publishVpnMetrics(ctx *zedrouterContext,
-	status *types.NetworkServiceStatus, vpnStatus *types.ServiceVpnStatus) {
-	metrics := new(types.NetworkServiceMetrics)
-	metrics.UUID = status.UUID
-	metrics.DisplayName = status.DisplayName
-	metrics.Type = status.Type
-
-	// get older metrics, if any
-	vpnMetrics := new(types.VpnMetrics)
-	oldMetrics := lookupNetworkServiceMetrics(ctx, metrics.Key())
-
-	// for cumulative stats, take the old metrics stats
-	// and add the difference to the metrics
-	if oldMetrics != nil && oldMetrics.VpnMetrics != nil {
-		vpnMetrics.DataStat.InPkts = oldMetrics.VpnMetrics.DataStat.InPkts
-		vpnMetrics.DataStat.OutPkts = oldMetrics.VpnMetrics.DataStat.OutPkts
-	}
-
-	publishVpnConnMetrics(ctx, status, oldMetrics, vpnMetrics, vpnStatus)
-	publishVpnMetricsAclCounters(vpnMetrics)
-	metrics.VpnMetrics = vpnMetrics
-	pub := ctx.pubNetworkServiceMetrics
-	pub.Publish(metrics.Key(), &metrics)
-	return
-}
-
 func publishVpnMetricsAclCounters(vpnMetrics *types.VpnMetrics) {
 
 	if pktStat, err := iptableCounterRuleStat(vpnCounterAcls[0]); err == nil {
@@ -899,62 +620,6 @@ func publishVpnMetricsAclCounters(vpnMetrics *types.VpnMetrics) {
 	if pktStat, err := iptableCounterRuleStat(vpnCounterAcls[5]); err == nil {
 		vpnMetrics.EspStat.OutPkts = pktStat
 	}
-}
-
-func publishVpnConnMetrics(ctx *zedrouterContext,
-	status *types.NetworkServiceStatus, oldMetrics *types.NetworkServiceMetrics,
-	vpnMetrics *types.VpnMetrics, vpnStatus *types.ServiceVpnStatus) {
-
-	if len(vpnStatus.ActiveVpnConns) == 0 {
-		return
-	}
-	vpnMetrics.VpnConns = make([]*types.VpnConnMetrics,
-		len(vpnStatus.ActiveVpnConns))
-	for idx, connStatus := range vpnStatus.ActiveVpnConns {
-		connMetrics := new(types.VpnConnMetrics)
-		connMetrics.Id = connStatus.Id
-		connMetrics.Name = connStatus.Name
-		connMetrics.Type = status.Type
-		connMetrics.EstTime = connStatus.EstTime
-		connMetrics.LEndPoint.IpAddr = connStatus.LInfo.IpAddr
-		connMetrics.REndPoint.IpAddr = connStatus.RInfo.IpAddr
-
-		// get the last metrics
-		oldConnMetrics := getVpnMetricsOldConnStats(oldMetrics, connStatus.Id)
-		// loop through the current setof SAs
-		for _, linkStatus := range connStatus.Links {
-			if linkStatus.State != types.VPN_INSTALLED {
-				continue
-			}
-			lStats := linkStatus.LInfo
-			connMetrics.LEndPoint.LinkInfo.SpiId = lStats.SpiId
-			connMetrics.LEndPoint.LinkInfo.SubNet = lStats.SubNet
-			connMetrics.LEndPoint.PktStats = lStats.PktStats
-
-			rStats := linkStatus.RInfo
-			connMetrics.REndPoint.LinkInfo.SpiId = rStats.SpiId
-			connMetrics.REndPoint.LinkInfo.SubNet = rStats.SubNet
-			connMetrics.REndPoint.PktStats = rStats.PktStats
-
-			// increment cumulative stats
-			incrementVpnMetricsConnStats(vpnMetrics,
-				oldConnMetrics, linkStatus)
-		}
-		vpnMetrics.VpnConns[idx] = connMetrics
-	}
-}
-
-func getVpnMetricsOldConnStats(oldMetrics *types.NetworkServiceMetrics,
-	id string) *types.VpnConnMetrics {
-	if oldMetrics == nil || oldMetrics.VpnMetrics == nil {
-		return nil
-	}
-	for _, connStatus := range oldMetrics.VpnMetrics.VpnConns {
-		if connStatus.Id == id {
-			return connStatus
-		}
-	}
-	return nil
 }
 
 // get the cumulative stats
